requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
mysqlclient==2.2.0
